% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCov.R
\name{getCov}
\alias{getCov}
\alias{mvrnorm2}
\title{Estimation Helper Functions}
\usage{
getCov(t1, t2, model, p)
}
\arguments{
\item{t1}{time 1}

\item{t2}{time 2}

\item{model}{the model}

\item{p}{vector of the auto-correlation parameters i.e. p = c(tau.z, tau.v)}
}
\description{
functions which provide the theoretical covariance \code{\link[=getCov]{getCov()}} and area \code{\link[=getArea]{getArea()}} for specific models and parameter values
}
\details{
\code{getCov(t1, t2, model, p)} calculates the covariance matrix for different models. \code{mvrnorm2} is a slightly more efficient multivariate normal function.
}
