% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCov.R
\name{getCov}
\alias{getArea,}
\alias{getCov}
\alias{getCov,}
\alias{mvrnorm2}
\title{Estimation Helper Functions}
\usage{
getCov(t1, t2, model, p)
}
\description{
functions which provide the theoretical covariance [getCov()] and area [getArea()] for specific models and parameter values
}
\details{
\code{getCov(t1, t2, model, p)} calculates the covariance matrix for different models. \code{mvrnorm2} is a slightly more efficient multivariate normal function.
}

